Handle underscores in type names
Whilst it's unusual to have an underscore in a type name, it's
legal in any identifier[1].

For syntax highlighting, we generally use the heuristic that an
identifier starting with an uppercase letter is probably a type.

Allow type names to start with an underscore before their uppercase
character, and allow underscores afterwards.

This fixes cases like:

```
struct _Foo_Bar {}
```

[1] https://doc.rust-lang.org/reference/identifiers.html